requirements-windows.txt
aiofiles>=24.0.0
aiohappyeyeballs>=2.6.0
aiohttp>=3.12.0
aiosignal>=1.4.0
annotated-types>=0.7.0
anyio>=4.9.0
attrs>=25.3.0
audioop-lts>=0.2.1
Brotli>=1.1.0
certifi>=2025.7.0
charset-normalizer>=3.4.0
click>=8.1.0
dataclasses-json>=0.6.0
distro>=1.9.0
fastapi>=0.116.0
ffmpy>=0.6.0
filelock>=3.18.0
frozenlist>=1.7.0
fsspec>=2025.7.0
gradio>=5.38.0
gradio_client>=1.11.0
groovy>=0.1.2
gTTS>=2.5.0
h11>=0.16.0
hf-xet>=1.1.0
httpcore>=1.0.0
httpx>=0.28.0
httpx-sse>=0.4.0
huggingface-hub>=0.33.0
idna>=3.10
Jinja2>=3.1.0
jiter>=0.10.0
jsonpatch>=1.33
jsonpointer>=3.0.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-ollama>=0.3.0
langchain-openai>=0.3.0
langchain-text-splitters>=0.3.0
langsmith>=0.4.0
markdown-it-py>=3.0.0
MarkupSafe>=3.0.0
marshmallow>=3.26.0
mdurl>=0.1.0
mpg123>=0.4
multidict>=6.6.0
mypy_extensions>=1.1.0
numpy>=2.3.0
ollama>=0.5.0
openai>=1.97.0
orjson>=3.11.0
packaging>=25.0
pandas>=2.3.0
pillow>=11.3.0
propcache>=0.3.0
PyAudio>=0.2.14
pydantic>=2.11.0
pydantic-settings>=2.10.0
pydantic_core>=2.33.0
pydub>=0.25.0
Pygments>=2.19.0
PyQt5>=5.15.0
python-dateutil>=2.9.0
python-dotenv>=1.1.0
python-multipart>=0.0.20
pyttsx3>=2.99
pytz>=2025.2
PyYAML>=6.0.0
regex>=2024.11.0
requests>=2.32.0
requests-toolbelt>=1.0.0
rich>=14.0.0
ruff>=0.12.0
safehttpx>=0.1.0
semantic-version>=2.10.0
shellingham>=1.5.0
six>=1.17.0
sniffio>=1.3.0
SpeechRecognition>=3.14.0
SQLAlchemy>=2.0.0
standard-aifc>=3.13.0
standard-chunk>=3.13.0
starlette>=0.47.0
tenacity>=9.1.0
tiktoken>=0.9.0
tomlkit>=0.13.0
tqdm>=4.67.0
typer>=0.16.0
typing-inspect>=0.9.0
typing-inspection>=0.4.0
typing_extensions>=4.14.0
tzdata>=2025.2
urllib3>=2.5.0
uvicorn>=0.35.0
websockets>=15.0.0
yarl>=1.20.0
zstandard>=0.23.0
